# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/asset_categories:
    get:
      tags:
        - AssetCategories
      summary: List AssetCategories
      description: Get a paginated list of AssetCategories with optional filtering and sorting
      operationId: listAssetCategories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCategoryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AssetCategories
      summary: Create AssetCategory
      description: Create a new AssetCategory
      operationId: createAssetCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetCategoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/asset_depreciation_entries:
    get:
      tags:
        - AssetDepreciationEntries
      summary: List AssetDepreciationEntries
      description: Get a paginated list of AssetDepreciationEntries with optional filtering and sorting
      operationId: listAssetDepreciationEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDepreciationEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AssetDepreciationEntries
      summary: Create AssetDepreciationEntry
      description: Create a new AssetDepreciationEntry
      operationId: createAssetDepreciationEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetDepreciationEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDepreciationEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    AssetCategory:
      type: object
      required:
        - id
        - company_id
        - category_name
        - depreciation_method
        - useful_life_months
        - fixed_asset_account_id
        - accumulated_depreciation_account_id
        - depreciation_expense_account_id
        - disposal_gain_loss_account_id
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        category_name:
          type: string
          maxLength: 140
        depreciation_method:
          type: string
        useful_life_months:
          type: integer
          format: int32
        fixed_asset_account_id:
          type: string
          format: uuid
        accumulated_depreciation_account_id:
          type: string
          format: uuid
        depreciation_expense_account_id:
          type: string
          format: uuid
        disposal_gain_loss_account_id:
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    CreateAssetCategoryRequest:
      type: object
      required:
        - company_id
        - category_name
        - depreciation_method
        - useful_life_months
        - fixed_asset_account_id
        - accumulated_depreciation_account_id
        - depreciation_expense_account_id
        - disposal_gain_loss_account_id
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        category_name:
          type: string
          maxLength: 140
        depreciation_method:
          type: string
        useful_life_months:
          type: integer
          format: int32
        fixed_asset_account_id:
          type: string
          format: uuid
        accumulated_depreciation_account_id:
          type: string
          format: uuid
        depreciation_expense_account_id:
          type: string
          format: uuid
        disposal_gain_loss_account_id:
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    AssetCategoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AssetCategory'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    AssetDepreciationEntry:
      type: object
      required:
        - id
        - company_id
        - asset_id
        - period_no
        - schedule_date
        - depreciation_amount
        - accumulated_after
        - posted
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        period_no:
          type: integer
          format: int32
        schedule_date:
          type: string
          format: date-time
        depreciation_amount:
          type: string
        accumulated_after:
          type: string
        posted:
          type: boolean
        posted_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateAssetDepreciationEntryRequest:
      type: object
      required:
        - company_id
        - asset_id
        - period_no
        - schedule_date
        - depreciation_amount
        - accumulated_after
        - posted
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        period_no:
          type: integer
          format: int32
        schedule_date:
          type: string
          format: date-time
        depreciation_amount:
          type: string
        accumulated_after:
          type: string
        posted:
          type: boolean
        posted_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    AssetDepreciationEntryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AssetDepreciationEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []