# src/exports/types.rs
//! Public types for Assets module
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! These DTOs are the ONLY types other modules should use.
//! They are decoupled from internal domain entities.

use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};
use rust_decimal::Decimal;
use crate::domain::entity::*;

// ============================================================================
// ASSETCATEGORY TYPES
// ============================================================================

/// Type-safe ID for AssetCategory
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct AssetCategoryId(pub Uuid);

impl AssetCategoryId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for AssetCategoryId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<AssetCategoryId> for Uuid {
    fn from(id: AssetCategoryId) -> Self {
        id.0
    }
}

/// Data transfer object for AssetCategory
///
/// This is the public representation of AssetCategory for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetCategoryDto {
    pub id: AssetCategoryId,
    pub company_id: Uuid,
    pub category_name: String,
    pub depreciation_method: DepreciationMethod,
    pub useful_life_months: i32,
    pub fixed_asset_account_id: Uuid,
    pub accumulated_depreciation_account_id: Uuid,
    pub depreciation_expense_account_id: Uuid,
    pub disposal_gain_loss_account_id: Uuid,
    pub status: AssetCategoryStatus,
    pub metadata: serde_json::Value,
}

/// Summary view of AssetCategory for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetCategorySummary {
    pub id: AssetCategoryId,
    pub category_name: String,
    pub status: AssetCategoryStatus,
}

/// Reference to AssetCategory for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetCategoryRef {
    pub id: AssetCategoryId,
}

// ============================================================================
// ASSET TYPES
// ============================================================================

/// Type-safe ID for Asset
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct AssetId(pub Uuid);

impl AssetId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for AssetId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<AssetId> for Uuid {
    fn from(id: AssetId) -> Self {
        id.0
    }
}

/// Data transfer object for Asset
///
/// This is the public representation of Asset for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetDto {
    pub id: AssetId,
    pub company_id: Uuid,
    pub asset_category_id: Uuid,
    pub asset_name: String,
    pub asset_code: String,
    pub item_id: Option<Uuid>,
    pub branch_id: Option<Uuid>,
    pub gross_purchase_amount: Decimal,
    pub salvage_value: Decimal,
    pub useful_life_months: i32,
    pub opening_accumulated_depreciation: Decimal,
    pub purchase_date: DateTime<Utc>,
    pub available_for_use_date: Option<DateTime<Utc>>,
    pub accumulated_depreciation: Decimal,
    pub net_book_value: Decimal,
    pub status: AssetStatus,
    pub metadata: serde_json::Value,
}

/// Summary view of Asset for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetSummary {
    pub id: AssetId,
    pub asset_name: String,
    pub status: AssetStatus,
}

/// Reference to Asset for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetRef {
    pub id: AssetId,
}

// ============================================================================
// ASSETDEPRECIATIONENTRY TYPES
// ============================================================================

/// Type-safe ID for AssetDepreciationEntry
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct AssetDepreciationEntryId(pub Uuid);

impl AssetDepreciationEntryId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for AssetDepreciationEntryId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<AssetDepreciationEntryId> for Uuid {
    fn from(id: AssetDepreciationEntryId) -> Self {
        id.0
    }
}

/// Data transfer object for AssetDepreciationEntry
///
/// This is the public representation of AssetDepreciationEntry for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetDepreciationEntryDto {
    pub id: AssetDepreciationEntryId,
    pub company_id: Uuid,
    pub asset_id: Uuid,
    pub period_no: i32,
    pub schedule_date: DateTime<Utc>,
    pub depreciation_amount: Decimal,
    pub accumulated_after: Decimal,
    pub posted: bool,
    pub posted_at: Option<DateTime<Utc>>,
    pub metadata: serde_json::Value,
}

/// Summary view of AssetDepreciationEntry for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetDepreciationEntrySummary {
    pub id: AssetDepreciationEntryId,
}

/// Reference to AssetDepreciationEntry for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AssetDepreciationEntryRef {
    pub id: AssetDepreciationEntryId,
}

// ============================================================================
// CUSTOM TYPES
// ============================================================================

// <<< CUSTOM TYPES START >>>
// Add custom public types here
// <<< CUSTOM TYPES END >>>